pyproject.toml
[project]
name = "lepidexbackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.14",
    "numpy>=2.0.2",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "tensorflow>=2.19.0",
    "uvicorn>=0.35.0",
]
